Day 09: implement part 1
* Use longs instead of integers for Intcode memory
* Output now returns a List instead of 1 value
* Add relative mode support
* Fix storing of input
* Allocated memory size can now be increased
* Upgrade Spring Boot to 2.4.0 (was 2.2.2)

+++ src/main/java/nl/smerik/adventofcode/aoc2019/controller/AdventOfCode2019Day02Controller.java
@@ -18,7 +18,7 @@ public class AdventOfCode2019Day02Controller {
     }
 
     @GetMapping("/part-01")
-    public ResponseEntity<Integer> getSolutionPart01() {
+    public ResponseEntity<Long> getSolutionPart01() {
         return ResponseEntity.ok(day02Service.getSolutionPart1());
     }
